Add files via upload
So in this project i have used a superstore sales data set from Kaggle. In this project i have solved some question related to to sales of the store for exploratory analysis. I have used libraries like Matplotlib, Pandas, Numpy and Plotly Express for visualization. I have also done sales forecasting in this also so that we can predict the sales. This is my first python project of data analysis